removed unwanted files and improved hooks
