+++ cucumber.json
@@ -8,7 +8,8 @@
         "dryRun": false,
         "require":[
             "src/test/steps/*.ts",
-            "src/test/steps/shared/*.ts"
+            "src/test/steps/shared/*.ts",
+            "src/test/support/*.ts"
         ],
         "requireModule":[
             "ts-node/register"
